requirements.txt
google-api-python-client==2.97.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
watchdog==3.0.0
python-dotenv==1.0.0
requests==2.31.0
runpod==1.3.0
pydrive2==1.15.4
tenacity==8.2.3
